Cleanup

+++ src/listeners/MessageReactionAdd.ts
@@ -1,8 +1,9 @@
 import { Events, GuildTextBasedChannel, MessageReaction, User } from "discord.js";
-import EventListener from "../handlers/listeners/EventListener";
-import ClientManager from "../Client";
 import { muteMember } from "../utils/ModerationUtils";
 import { RolePermission } from "../utils/Types";
+
+import EventListener from "../handlers/listeners/EventListener";
+import ClientManager from "../Client";
 
 export default class MessageReactionAddEventListener extends EventListener {
     constructor() {
@@ -13,58 +14,62 @@ export default class MessageReactionAddEventListener extends EventListener {
     }
 
     async execute(reaction: MessageReaction, user: User): Promise<void> {
+        const { emoji, message } = reaction;
+
+        if (user.bot || !message.guild) return;
         if (reaction.partial) await reaction.fetch();
-        if (user.bot) return;
 
         const config = ClientManager.config(reaction.message.guildId!)!;
-        const member = await reaction.message.guild!.members.fetch(user.id);
+        const member = await message.guild.members.fetch(user.id);
 
         if (!config.isGuildStaff(member)) return;
 
-        const { emoji, message } = reaction;
-        const emojiName = emoji.id ?? emoji.name ?? "N/A";
+        const emojiId = emoji.id ?? emoji.name ?? "N/A";
 
-        // Quick mutes
-        if (config.emojis.quickMute30?.includes(emojiName) || config.emojis.quickMute60?.includes(emojiName)) {
+        /* Quick mutes - 30 minutes and 1 hour */
+        if (config.emojis.quickMute30?.includes(emojiId) || config.emojis.quickMute60?.includes(emojiId)) {
             if (!message.member) return;
 
-            let duration = "30m";
-            let emojiType = "quickMute30";
+            let muteDuration = "30m";
+            let emojiName = "quickMute30";
 
-            if (config.emojis.quickMute60?.includes(emojiName)) {
-                duration = "60m";
-                emojiType = "quickMute60";
+            if (config.emojis.quickMute60?.includes(emojiId)) {
+                muteDuration = "60m";
+                emojiName = "quickMute60";
             }
 
             if (!config.actionAllowed(member, {
-                property: RolePermission.Reaction,
-                value: emojiType
+                permission: RolePermission.Reaction,
+                requiredValue: emojiName
             })) return;
 
             const reason = message.content ?? undefined;
-            const res = await muteMember({
-                offender: message.member,
+            const res = await muteMember(message.member, {
                 moderator: user,
-                duration,
+                duration: muteDuration,
                 config,
                 reason
             });
 
-            const channelId = config.channels.staffCommands;
-            if (!channelId) return;
+            const confirmationChannelId = config.channels.staffCommands;
+            if (!confirmationChannelId) return;
 
-            const channel = await message.guild?.channels.fetch(channelId) as GuildTextBasedChannel;
-            if (!channel) return;
+            const confirmationChannel = await message.guild.channels.fetch(confirmationChannelId) as GuildTextBasedChannel;
+            if (!confirmationChannel) return;
 
             await reaction.remove();
 
+            /* The result is the mute's expiration timestamp */
             if (typeof res === "number") {
-                await channel.send(`${config.emojis.success} **${user.tag}** has successfully muted **${message.author?.tag}** until <t:${res}:F> | Expires <t:${res}:R> (\`${reason}\`)`);
-                await message.delete().catch(() => null);
+                await Promise.all([
+                    confirmationChannel.send(`${config.emojis.success} **${user.tag}** has successfully muted **${message.author?.tag}** until <t:${res}:F> | Expires <t:${res}:R> (\`${reason}\`)`),
+                    message.delete().catch(() => null)
+                ]);
                 return;
             }
 
-            await channel.send(`${config.emojis.error} ${user} ${res}`);
+            /* The result is an error message */
+            await confirmationChannel.send(`${config.emojis.error} ${user} ${res}`);
         }
     }
 }
